删除子表

+++ icity-iform-service/src/main/java/tech/ascs/icity/iform/service/impl/FormModelServiceImpl.java
@@ -726,7 +726,9 @@ public class FormModelServiceImpl extends DefaultJPAService<FormModelEntity> imp
 				}
 				itemManager.delete(itemModelEntity);
 			}else if(itemModelEntity instanceof SubFormItemModelEntity){
-				columnModelService.deleteTable(itemModelEntity.getColumnModel().getDataModel().getTableName());
+				if(itemModelEntity.getColumnModel() != null) {//刪除子表
+					columnModelService.deleteTable(itemModelEntity.getColumnModel().getDataModel().getTableName());
+				}
 				for(int t =0 ; t < ((SubFormItemModelEntity) itemModelEntity).getItems().size(); t++) {
 					SubFormRowItemModelEntity subFormRowItemModelEntity = ((SubFormItemModelEntity) itemModelEntity).getItems().get(t);
 					List<ItemModelEntity> list1 = subFormRowItemModelEntity.getItems();
